backend/requirements.txt
fastapi
uvicorn
alt-profanity-check
pyjwt
twilio
psycopg2-binary
sqlalchemy
sqlalchemy-cockroachdb
python-decouple
